// web/src/api/wire.ts
export type ExerciseType =
  | "Bodyweight"
  | "WeightedBodyweight"
  | "AssistedBodyweight"
  | "Weighted";
// ...
export type MuscleGroup =
  | "Chest"
  | "Back"
  | "Traps"
  | "Shoulders"
  | "Quads"
  | "Hamstrings"
  | "Glutes"
  | "Biceps"
  | "Triceps"
  | "Calves";

/**
 * Input enum mirroring [`MuscleGroup`].
 */
export type MuscleGroupInput =
  | "Chest"
  | "Back"
  | "Traps"
  | "Shoulders"
  | "Quads"
  | "Hamstrings"
  | "Glutes"
  | "Biceps"
  | "Triceps"
  | "Calves";

export type Name = string;

export type PercentOneRepMax = number;

export type Phase = "Accumulation" | "Intensification" | "Deload";